Move ArraySource to separate file
